requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
pydantic-settings
pyautogen
sqlalchemy
psychopg2